+++ R/hugo.R
@@ -565,6 +565,15 @@ content_file = function(...) file.path(
 #' @param categories A character vector of category names.
 #' @param tags A character vector of tag names.
 #' @param date The date of the post.
+#' @param time Whether to include the time of the day in the \code{date} field
+#'   of the post. If \code{TRUE}, the \code{date} will be of the format
+#'   \samp{\%Y-\%m-\%dT\%H:\%M:\%S\%z} (e.g., \samp{2001-02-03T04:05:06-0700}).
+#'   Alternatively, it can take a character string to be appended to the
+#'   \code{date}. It can be important and helpful to include the time in the
+#'   date of a post. For example, if your website is built on a server (such as
+#'   Netlify or Vercel) and your local timezone is ahead of UTC, your local date
+#'   may be a \emph{future} date on the server, and Hugo will not build future
+#'   posts by default (unless you use the \command{-F} flag).
 #' @param file The filename of the post. By default, the filename will be
 #'   automatically generated from the title by replacing non-alphanumeric
 #'   characters with dashes, e.g. \code{title = 'Hello World'} may create a file
@@ -593,8 +602,8 @@ content_file = function(...) file.path(
 #'   that the author field is automatically filled out when creating a new post.
 new_post = function(
   title, kind = '', open = interactive(), author = getOption('blogdown.author'),
-  categories = NULL, tags = NULL, date = Sys.Date(), file = NULL, slug = NULL,
-  title_case = getOption('blogdown.title_case'),
+  categories = NULL, tags = NULL, date = Sys.Date(), time = getOption('blogdown.time', FALSE),
+  file = NULL, slug = NULL, title_case = getOption('blogdown.title_case'),
   subdir = getOption('blogdown.subdir', 'post'), ext = getOption('blogdown.ext', '.md')
 ) {
   if (is.null(file)) file = post_filename(title, subdir, ext, date)
@@ -627,8 +636,8 @@ new_post = function(
   }
 
   do.call(modify_yaml, c(list(
-    file, title = title, author = author, date = format(date), slug = slug,
-    categories = modify_field(categories), tags = modify_field(tags)
+    file, title = title, author = author, date = format_datetime(date, time),
+    slug = slug, categories = modify_field(categories), tags = modify_field(tags)
   ), if (kind == '' && !file.exists('archetypes/default.md')) list(draft = NULL)
   ))
   open_file(file, open)